End finished sessions on timeline updates.
We currently only end sessions on Timeline updates if the associated media
is no longer in the playlist. But we should also end all sessions that are
finished as a result of the timeline update (similar to how this is done for
discontinuities). This issue was introduced by https://github.com/google/ExoPlayer/commit/897aa91f62a47cc92cfa1b0513fd6edae977b385

PiperOrigin-RevId: 423075855